Rename personal_exceprions to custom_exceptions, fix bug with --json and --date, redo --to-fb2 and --to-html: now the argument is directory and file news.fb2 or news.html are created in this directory. Some other little changes
